(refactor)[Sun 04/16/2023] refactor and small fixes
