--- src/components/explore/curve.css
@@ -1,21 +1,0 @@
-.custom-shape-divider-top-1680108220 {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 100%;
-    overflow: hidden;
-    line-height: 0;
-    transform: rotate(180deg);
-}
-
-.custom-shape-divider-top-1680108220 svg {
-    position: relative;
-    display: block;
-    width: calc(100% + 1.3px);
-    height: 239px;
-    transform: rotateY(180deg);
-}
-
-.custom-shape-divider-top-1680108220 .shape-fill {
-    fill:gray;
-}
