--- src/components/explore/desktop/Scroll.tsx
@@ -1,79 +1,0 @@
-import React from 'react';
-
-const Scroll = () => {
-  return (
-    <div className='flex justify-center items-center mt-12'>
-      <button type='button'>
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          fill='none'
-          viewBox='0 0 24 24'
-          strokeWidth='1.5'
-          stroke='currentColor'
-          className='w-4 h-4'
-        >
-          <path
-            strokeLinecap='round'
-            strokeLinejoin='round'
-            d='M18.75 19.5l-7.5-7.5 7.5-7.5m-6 15L5.25 12l7.5-7.5'
-          />
-        </svg>
-      </button>
-
-      <button type='button'>
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          fill='none'
-          viewBox='0 0 24 24'
-          strokeWidth={1.5}
-          stroke='currentColor'
-          className='w-5 h-5'
-        >
-          <path
-            strokeLinecap='round'
-            strokeLinejoin='round'
-            d='M15.75 19.5L8.25 12l7.5-7.5'
-          />
-        </svg>
-      </button>
-
-      <span className='font-bold text-lg px-2'>1</span>
-
-      <button type='button'>
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          fill='none'
-          viewBox='0 0 24 24'
-          strokeWidth={1.5}
-          stroke='currentColor'
-          className='w-5 h-5'
-        >
-          <path
-            strokeLinecap='round'
-            strokeLinejoin='round'
-            d='M8.25 4.5l7.5 7.5-7.5 7.5'
-          />
-        </svg>
-      </button>
-
-      <button type='button'>
-        <svg
-          xmlns='http://www.w3.org/2000/svg'
-          fill='none'
-          viewBox='0 0 24 24'
-          strokeWidth={1.5}
-          stroke='currentColor'
-          className='w-4 h-4'
-        >
-          <path
-            strokeLinecap='round'
-            strokeLinejoin='round'
-            d='M11.25 4.5l7.5 7.5-7.5 7.5m-6-15l7.5 7.5-7.5 7.5'
-          />
-        </svg>
-      </button>
-    </div>
-  );
-};
-
-export default Scroll;
