--- src/components/explore/roadmapTests.json
@@ -1,65 +1,0 @@
-[
-    {
-        "id": 1,
-        "name": "React",
-        "madeby": "RusBoss",
-        "nolikes": 23000,
-        "description": "Really cool react tooling with fast algorithms and skill improvement"
-    },
-    {
-        "id": 2,
-        "name": "Typescript",
-        "madeby": "RusBoss",
-        "nolikes": 42000,
-        "description": "Learn a new javascript, but with steroids."
-    },
-    {
-        "id": 3,
-        "name": "Javascript",
-        "madeby": "RusBoss",
-        "nolikes": 76000,
-        "description": "Everyone knows javascript come on.."
-    },
-    {
-        "id": 4,
-        "name": "b",
-        "madeby": "RusBoss",
-        "nolikes": 76000,
-        "description": "Everyone knows javascript come on.."
-    },
-    {
-        "id": 5,
-        "name": "c",
-        "madeby": "RusBoss",
-        "nolikes": 76000,
-        "description": "Everyone knows javascript come on.."
-    },
-    {
-        "id": 6,
-        "name": "d",
-        "madeby": "RusBoss",
-        "nolikes": 76000,
-        "description": "Everyone knows javascript come on.."
-    },
-    {
-        "id": 7,
-        "name": "e",
-        "madeby": "RusBoss",
-        "nolikes": 76000,
-        "description": "Everyone knows javascript come on.."
-    },
-    {
-        "id": 8,
-        "name": "f",
-        "madeby": "RusBoss",
-        "nolikes": 76000,
-        "description": "Everyone knows javascript come on.."
-    },
-    {
-        "id": 9,
-        "name": "g",
-        "madeby": "RusBoss",
-        "nolikes": 76000,
-        "description": "Everyone knows javascript come on.."
-    }
-]
