+++ src/pages/explore/index.astro
@@ -1,27 +1,31 @@
 ---
 import Base from '@layouts/Base.astro';
-import SearchMobile from '@components/explorerefr/mobile/Search';
+import SearchMobile from '../../components/explorerefr/mobile/Explore';
 import SearchDesktop from '@components/explorerefr/desktop/Search';
-import wave from '@assets/wavesNegative.svg'
-// import TestDawg from '@components/explorerefr/ExampleHOC/TestDawg';
+import wave from '@assets/wavesNegative.svg';
 ---
 
 <Base title='Welcome to Astro.'>
-  <div class="hidden md:block">
-    <div class="mt-40 relative">
-      <div class='bg-[#cfeefc] absolute w-full h-[300px] top-[-160px] -z-20' />
-      <img src={wave} alt="WaveBackground" class="w-full absolute -z-10">
-      <div class="flex justify-center items-center">
-        <h1 class="text-4xl font-kanit-text mt-2">The journey of 1000 miles starts with 1 roadmap</h1>
+  <div class='hidden lg:block'>
+    <div class='mt-40 relative'>
+      <div class='bg-[#cfeefc] absolute w-full h-[300px] top-[-160px] -z-20'>
       </div>
-      <SearchDesktop client:load/>
+      <img src={wave} alt='WaveBackground' class='w-full absolute -z-10' />
+      <div class='flex justify-center items-center'>
+        <h1 class='text-4xl font-semibold font-kanit-text mt-2'>
+          The journey of 1000 miles starts with 1 roadmap
+        </h1>
+      </div>
+      <SearchDesktop client:idle />
     </div>
   </div>
-  <div class="md:hidden">
-    <div class="flex items-center justify-center font-kanit-text text-[21px] flex-col mt-9 sm:text-[31px]">
+  <div class='lg:hidden'>
+    <div
+      class='flex items-center justify-center font-kanit-text text-[21px] flex-col mt-9 sm:text-[31px]'
+    >
       <h1>The journey of 1000 steps</h1>
       <h1>starts with 1 roadmap</h1>
     </div>
-    <SearchMobile client:load/>
+    <SearchMobile client:idle />
   </div>
 </Base>
